requirements.txt
openai>=1.27.0
playwright>=1.44.0
python-dotenv>=1.0.1
